Made Collapsible itinerary a component
This is to set up for mapping events.
Important to note that the mapping of the itineraries will be by DAY
